3sum success submitted
Key facts:
How to solve the two sums.

Previously I use the hash table to solve the two sums, so I didn't think there is a way could solve which is use two pointer. The idea is,
1. Sort the array.
2. Use two point, left from array left and right from array right. If left + right > lookingfor, then it means right is too big, then move right to -1 position and retry. If it's small than looking for means left is small, so move left +1 and retry.
If we found the lookingFor, to avoid duplicates. Since the array is sorted, so we should adjust the left, right to move all duplicates by compare it with its neighbors.

For three sum, basically is search for each number N in the array and looking for -N. The same problem is avoid the duplicate which same idea is don't search for numbers we already searched. This is achieved by
1. If the neighbor of N( N + 1) is same as N, then it's already searched, then continue skip until a new one.
2. For each N, we only do search for rest of N, for example if the index is K, then we only search for two sum in (K + 1 .. Size of array -1 )