searchMatrix *
Found a good StackOverflow answer. The referenced paper by Richard Bird should be a good read if you can get it. (http://stackoverflow.com/questions/2457792/given-a-2d-array-sorted-in-increasing-order-from-left-to-right-and-top-to-bottom/2458113#2458113)

The referenced article by Craig Gidney nicely explains the Ω(m * log(M / m)) lower bound (where m is the smaller and M is the larger of height and width, and log is the normal logarithm except it doesn't go below 1). So the answer to your question is: No.(http://twistedoakstudios.com/blog/Post5365_searching-a-sorted-matrix-faster)

There's also a good LeetCode article about this problem with several solutions, though the complexity analysis isn't complete. (http://articles.leetcode.com/searching-2d-sorted-matrix-part-ii)